delete progress_bar module and place functions in general.py
